Update CLI tests for organization
